+++ app/src/main/kotlin/de/ids_mannheim/korapxmltools/KorapXmlTool.kt
@@ -61,7 +61,7 @@ val ZIP_ENTRY_UNIX_MODE = parseInt("644", 8)
 @Command(
     name = "korapxmltool",
     mixinStandardHelpOptions = true,
-    version = ["korapxmltool v3.1.1"],
+    version = ["korapxmltool v3.1.2"],
     usageHelpAutoWidth = false,
     usageHelpWidth = 200,
     description = ["Converts between KorAP-XML ZIP format and formats like CoNLL-U, Krill, word2vec, NOW\n"+